minor issues

+++ projects/AdminUI/master/js/modules/controllers/users.js
@@ -3,15 +3,15 @@
  *smangipudi
  =========================================================*/
 
-App.controller('UsersController', ['$scope', '$state','$compile','$timeout', 'RestServiceFactory','DataTableService', 'toaster','UserRoleService',
-                                   function($scope, $state, $compile, $timeout, RestServiceFactory, DataTableService, toaster, UserRoleService) {
+App.controller('UsersController', ['$scope', '$state','$compile','$timeout', 'RestServiceFactory','DataTableService', 'toaster','UserRoleService','DialogService',
+                           function($scope, $state, $compile, $timeout, RestServiceFactory, DataTableService, toaster, UserRoleService, DialogService) {
   'use strict';
 
    	var userRoles = UserRoleService.getRoles();
   	$timeout(function(){
 
-    if ( ! $.fn.dataTable ) return;
-	    var columnDefinitions = [
+	    if ( ! $.fn.dataTable ) return;
+		var columnDefinitions = [
 	        { sWidth: "5%", aTargets: [0] },
 	        { sWidth: "15%", aTargets: [1,2,3] },
 	        { sWidth: "10%", aTargets: [4,5] },
@@ -31,7 +31,7 @@ App.controller('UsersController', ['$scope', '$state','$compile','$timeout', 'Re
 		    		imgHtml	+= '</div>';
 		    		$(td).html(imgHtml);
 		    		$compile(td)($scope);
-		    	  }
+	    	  	}
 	        },
 	        {
 		    	"targets": [6],
@@ -42,15 +42,14 @@ App.controller('UsersController', ['$scope', '$state','$compile','$timeout', 'Re
 		    		actionHtml += '<button title="Associate Venue" class="btn btn-default btn-oval fa fa-home"'+
 		    			' ng-click="associateVenue(' +row +','+cellData+')"></button>';
 		    		if (rowData[6] !== 1) {
-		    			actionHtml += '<button title="Delete User" class="btn btn-default btn-oval fa fa-trash"'+
-		    			' ng-click="deleteUser(' +row +','+cellData+')"></button>';
+		    			actionHtml += '<button title="Delete User" class="btn btn-default btn-oval fa fa-trash"></button>';
 		    		}
 		    		
 		    		$(td).html(actionHtml);
 		    		$compile(td)($scope);
-		    	  }
-	    	},
-	    	{
+			    }
+		    },
+		    {
 		    	"targets": [4],
 		    	"orderable": false,
 		    	"createdCell": function (td, cellData, rowData, row, col) {
@@ -62,58 +61,65 @@ App.controller('UsersController', ['$scope', '$state','$compile','$timeout', 'Re
 		    		$(td).html(actionHtml);
 		    		$compile(td)($scope);
 		    	}
-		 } ];
-    
-	    DataTableService.initDataTable('users_table', columnDefinitions);
-   
+		 	} 
+		];
+
+		DataTableService.initDataTable('users_table', columnDefinitions);
+
 	    var promise = RestServiceFactory.UserService().get();
 	    promise.$promise.then(function(data) {
-    	 
-    	var table = $('#users_table').DataTable();
-    	
-    	data.users.map(function(user) {
-    		var role = userRoles[user.roleId];
+	    	var table = $('#users_table').DataTable();
+	    	data.users.map(function(user) {
+	    		var role = userRoles[user.roleId];
 
-    		if (role == null) {
-    			role = user.role;
-    		}
-    		if (user.businessName == null) {
-    			user.businessName = "";
-    		}
-    		var img = user.profileImageThumbnail;
-    		if (typeof img === 'undefined') {
-    			img = '';
-    		}
-    		table.row.add([img, user.userName, user.loginId, user.businessName, user.enabled, role, user.id]);
-    	});
-    	table.draw();
-    });
-	    
-  	
-  });
-  
-  $scope.editUser = function(userId) {
+	    		if (role == null) {
+	    			role = user.role;
+	    		}
+	    		if (user.businessName == null) {
+	    			user.businessName = "";
+	    		}
+	    		var img = user.profileImageThumbnail;
+	    		if (typeof img === 'undefined') {
+	    			img = '';
+	    		}
+	    		table.row.add([img, user.userName, user.loginId, user.businessName, user.enabled, role, user.id]);
+	    	});
+			table.draw();
+		});
+		var table = $('#users_table').DataTable();
+		$('#users_table').on('click', '.fa-trash',function() {
+	  		$scope.deleteUser(this, table);
+	  	});
+	});
+	
+ 
+	$scope.editUser = function(userId) {
 		$state.go('app.useredit', {id: userId});
-  };
-  
-  $scope.associateVenue = function(rowId, userId) {
-		$state.go('app.uservenues', {id: userId});
-  };
-  
-  
-  $scope.deleteUser = function(rowId, userId) {	
-		var target = {id: userId};
-		RestServiceFactory.UserService().delete(target,  function(success){
-  		var table = $('#users_table').dataTable();
-  		table.fnDeleteRow(rowId);
-  	},function(error){
-  		if (typeof error.data !== 'undefined') { 
-  			toaster.pop('error', "Server Error", error.data.developerMessage);
-  		}
-  	});
 	};
+
+	$scope.associateVenue = function(rowId, userId) {
+		$state.go('app.uservenues', {id: userId});
+	};
+
   
-  $scope.createNewUser = function() {
+ 	$scope.deleteUser = function(button, table) {	
+	  	DialogService.confirmYesNo('Delete User?', 'Are you sure want to delete selected User?', function() {
+	        var targetRow = $(button).closest("tr");
+	        var rowData = table.row( targetRow).data();
+	 		var target = {id: rowData[6]};
+			RestServiceFactory.UserService().delete(target,  function(success){
+		  		table.row(targetRow).remove().draw();
+		  	},function(error){
+		  		if (typeof error.data !== 'undefined') { 
+		  			toaster.pop('error', "Server Error", error.data.developerMessage);
+		  		}
+	  		});
+	        
+   		 });
+
+ 	};
+  
+  	$scope.createNewUser = function() {
 		$state.go('app.useredit', {id: 'new'});
 	};
 }]);
